[UPDATE] HomeFragment.kt

+++ app/src/main/java/com/allana/food_recipe_app_chapter7/ui/features/favorite/FavoriteRecipeFragment.kt
@@ -34,7 +34,7 @@ FavoriteRecipeContract.View {
 
     override fun initList() {
         adapter = FavoriteRecipeAdapter {
-            DetailActivity.startActivity(context, it.idRecipe?.toInt() ?: 0)
+            DetailActivity.startActivity(context, it.idRecipe?.toLong() ?: 0)
         }
         getViewBinding().rvFavoriteRecipe.apply {
             layoutManager = LinearLayoutManager(context)
